requirements.txt
aiohappyeyeballs==2.4.3
aiohttp==3.10.10
aiosignal==1.3.1
anyio==4.6.2.post1
async-timeout==4.0.3
attrs==24.2.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
discord.py==2.4.0
exceptiongroup==1.2.2
frozenlist==1.5.0
h11==0.14.0
httpcore==1.0.7
httpx==0.27.2
idna==3.10
multidict==6.1.0
pillow==11.0.0
propcache==0.2.0
pycparser==2.22
PyNaCl==1.5.0
python-dotenv==1.0.1
pytube==15.0.0
redis==5.2.0
requests==2.32.3
sniffio==1.3.1
spotipy==2.24.0
typing_extensions==4.12.2
urllib3==2.2.3
yarl==1.17.1
youtube-search-python==1.6.6
yt-dlp==2024.11.4
